requirements.txt
# Reminder backend — pin versions in production
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pyyaml>=6.0.1,<7.0.0





# exact working verions:
# annotated-doc==0.0.4
# annotated-types==0.7.0
# anyio==4.12.1
# click==8.3.1
# colorama==0.4.6
# fastapi==0.135.1
# h11==0.16.0
# httptools==0.7.1
# idna==3.11
# pydantic==2.12.5
# pydantic_core==2.41.5
# python-dotenv==1.2.2
# PyYAML==6.0.3
# starlette==0.52.1
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# uvicorn==0.42.0
# watchfiles==1.1.1
# websockets==16.0